move getCurrentUser function to own file in utils
